65 º - Implementacion correcta de politicas y tests
Se han implementado de forma correcta las politicas, se ha quitado del app service provider el registro de las politicas ya que no es necesario (como tienen el mismo nombre que el modelo, laravel automaticamente hace lo suyo) y se han implementado en los controladores

+++ app/Http/Controllers/CategoryController.php
@@ -5,19 +5,24 @@ namespace App\Http\Controllers;
 use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use App\Models\Product;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class CategoryController extends Controller
 {
+    use AuthorizesRequests;
 
     public function index(){
+        $this->authorize('viewAny', Category::class);
         $categories = Category::all();
         return view('category.index', compact('categories'));
     }
     public function create(){
+        $this->authorize('create', Category::class);
         return view('category.create');
     }
 
     public function store(CategoryRequest $request){
+        $this->authorize('create', Category::class);
         if ($request->hasFile('icon')) {
             $path = $request->file('icon')->store('icons', 'public');
             $validated = $request->validated();
@@ -36,14 +41,19 @@ class CategoryController extends Controller
         return view('category.show', compact('category','products', 'categories'));
     }
     public function destroy(Category $category){
+        $this->authorize('delete', $category);
         $category->delete();
         return redirect()->route('categories')->with('status', __('Categoria eliminada'));
     }
 
     public function edit(Category $category){
+        $this->authorize('update', $category);
         return view('category.edit', compact('category'));
     }
     public function update(CategoryRequest $request, Category $category){
+
+        $this->authorize('update', $category);
+
         if ($request->hasFile('icon')) {
             $path = $request->file('icon')->store('icons', 'public');
             $validated = $request->validated();
