+++ app/Http/Controllers/ProductController.php
@@ -8,15 +8,18 @@ use App\Models\Category;
 use App\Models\Product;
 use App\Services\ProductImageService;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ProductController extends Controller
 {
-
+    use AuthorizesRequests;
     public function create() {
+        $this->authorize('create', Product::class);
         $categories = Category::all();
         return view('product.product-create', compact('categories'));
     }
     public function store(ProductRequest $request) {
+        $this->authorize('create', Product::class);
         $validated = $request->validated();
         $product = new Product($validated);
         $product->price = $validated['price'] * 100;
@@ -37,13 +40,16 @@ class ProductController extends Controller
         return redirect()->route('product.show', $product)->with('status', __('Producto creado!'));
     }
     public function show(Product $product) {
+
         return view('product.product-show', compact('product'));
     }
     public function edit(Product $product) {
+        $this->authorize('update', $product);
         $categories = Category::all();
         return view('product.product-edit', compact('product', 'categories'));
     }
     public function update(ProductUpdateRequest $request, Product $product) {
+        $this->authorize('update', $product);
 
         $validated = $request->validated();
 
@@ -67,12 +73,14 @@ class ProductController extends Controller
         return redirect()->route('product.show', $product)->with('status', __('Producto actualizado correctamente'));
     }
     public function destroy(Product $product) {
+        $this->authorize('delete', $product);
 
         $product->delete();
         return redirect()->route('index')->with('status', __('El producto ha sido eliminado correctamente'));
     }
 
     public function post(Product $product) {
+        $this->authorize('update', $product);
         $product->status = 'published';
         $product->save();
         return redirect()->route('product.show', $product)->with('status', __('Producto publicado correctamente'));
